Update Create Page with Handlers and Map

+++ src/main/java/edu/brown/cs/drawbridge/carpools/TripSearcher.java
@@ -141,7 +141,8 @@ public class TripSearcher {
    */
   public List<List<Trip>> searchWithId(String userId, double startLat,
       double startLon, double endLat, double endLon, long departureTime,
-      double distanceRadius, long timeRadius) {
+      double distanceRadius, double timeRadius) {
+    
     List<List<Trip>> paths = search(userId, startLat, startLon, endLat, endLon,
         departureTime, distanceRadius, timeRadius);
     Collections.sort(paths, TRIP_COMPARATOR);
@@ -172,7 +173,8 @@ public class TripSearcher {
    */
   public List<List<Trip>> searchWithoutId(double startLat, double startLon,
       double endLat, double endLon, long departureTime, double distanceRadius,
-      int timeRadius) {
+      double timeRadius) {
+
     List<List<Trip>> paths = search("", startLat, startLon, endLat, endLon,
         departureTime, distanceRadius, timeRadius);
     Collections.sort(paths, new CostComparator());
@@ -205,7 +207,8 @@ public class TripSearcher {
    */
   private List<List<Trip>> search(String userId, double startLat,
       double startLon, double endLat, double endLon, long departureTime,
-      double distanceRadius, long timeRadius) {
+      double distanceRadius, double timeRadius) {
+
     // Set the user id for all comparators
     setUser(userId);
